rename common module
