+++ .fossa.yml
@@ -351,7 +351,7 @@ targets:
       target: ':instrumentation:aws-lambda:aws-lambda-events-3.11:library'
     - type: gradle
       path: ./
-      target: ':instrumentation:aws-lambda:aws-lambda-events-common:library'
+      target: ':instrumentation:aws-lambda:aws-lambda-events-common-2.2:library'
     - type: gradle
       path: ./
       target: ':instrumentation:aws-sdk:aws-sdk-1.11:javaagent'
