+++ settings.gradle.kts
@@ -165,7 +165,7 @@ include(":instrumentation:aws-lambda:aws-lambda-events-2.2:javaagent")
 include(":instrumentation:aws-lambda:aws-lambda-events-2.2:library")
 include(":instrumentation:aws-lambda:aws-lambda-events-2.2:testing")
 include(":instrumentation:aws-lambda:aws-lambda-events-3.11:library")
-include(":instrumentation:aws-lambda:aws-lambda-events-common:library")
+include(":instrumentation:aws-lambda:aws-lambda-events-common-2.2:library")
 include(":instrumentation:aws-sdk:aws-sdk-1.11:javaagent")
 include(":instrumentation:aws-sdk:aws-sdk-1.11:library")
 include(":instrumentation:aws-sdk:aws-sdk-1.11:library-autoconfigure")
